refactor: updated tabs component to make it as more developer friendly
