+++ src/components/items/index.ts
@@ -1,0 +1,1 @@
+export * from './collection-items';
