+++ src/locales/en/translation.json
@@ -171,5 +171,9 @@
   },
   "nftCard": {
     "owned": "Owned"
+  },
+  "tabs": {
+    "items": "Items",
+    "activity": "Activity"
   }
 }
